202201081646FixAndUpdate 问题修复： 1.修复了Device类重复定义和头文件循环引用导致编译不通过的问题。 系统调整： 1.部分包含在连接组件中的业务模块被移动到了Core中。 2.为了与外部模块相兼容，使用的python版本降低至3.9.9。 3.优化了Control执行时钟逻辑，减少性能浪费。 功能更新： 1.添加了1.2.0R版本的前置更新内容。（1.2.0R即将上线的新特性：自适应线程调度、Python控制脚本支持，脚本代码热更新，CMD命令行等）

+++ MIBMSCloudSDK/private/MethodsLibrary.cpp
@@ -81,9 +81,9 @@ void VectorElementDelete_TaskInfo(SOCKET TClientSocket, vector<RecvTaskInfo> Tas
 	}
 }
 
-void Delay(int time)//time为秒数 
+void Delay(float time)//time为秒数 
 {
-	int T = time * 1000;
+	float T = time * 1000;
 	clock_t now = clock();
 	while (clock() - now < T);
 }
